+++ CHANGE_LOG.md
@@ -42,3 +42,5 @@ All notable changes to this project will be documented in this file.## [Unreleas
 - Updating widget specific `BroadcastService`
 - Updated props for auth chat
 - Adding `WidgetInstanceId` to support same widget with multi origin
+- Fixed custom context when `skipchatbuttonrendering` is enabled
+- Added exception for authenticated chat when context is passed